Do not sendError(500) when evaluating a template raises an exception
The rationale is that Pebble allows to flush template data at
developer's will, and trying to sendError after the response has been
written makes for a weirder and uglier exception. Logging the error and
just closing the connection creates a visually poorer experience for the
end user, but there's no solution that works 100% in all situations.